Working on getting mimic and diary commands
The biggest change is setting QoS to 1 which matches the door controller
arduino code.

Update status window to use the fail on health check option.

Update diary to listen to the right topics and update the time stamp for
the client each time it receives a message from that client.

Update topics so that the last level is the client id. This way it can
know which client the message originated from. Otherwise it would have
to be placed into the payload which isn't ideal imo.