Improvements to access_list and mimic
Improve error handling for the access_list command so that it returns an
exit code when it fails to do something critical. Also removed the await
connection call as it's not needed for this case.

Add access_list handlers to the mimic command that will publish fatal
commands when the access list option is toggled.